requirements.txt
alabaster==0.7.12
antlr4-python3-runtime==4.8
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
DataProperty==0.50.1
docutils==0.16
Faker==8.10.1
future==0.18.2
idna==2.10
imagesize==1.2.0
Jinja2==3.0.1
lark-parser==0.11.3
livereload==2.6.3
markdown-it-py==1.1.0
MarkupSafe==2.0.1
mbstrdecoder==1.0.1
mdit-py-plugins==0.2.8
myst-parser==0.15.1
packaging==20.9
pathvalidate==2.4.1
Pygments==2.9.0
pyparsing==2.4.7
pytablewriter==0.61.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.3
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinx-autobuild==2021.3.14
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-confluencebuilder==1.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stix2==2.1.0
stix2-generator==0.1.0
stix2-patterns==1.3.2
tabledata==1.1.4
tcolorpy==0.1.1
text-unidecode==1.3
tornado==6.1
typepy==1.1.5
typing-extensions==3.10.0.0
urllib3==1.26.5
